Move browser detection out of AppPage component
Increases separation of concerns and fixes bug where Edge browser only
blocks the page until the GraphQL response comes back.
