+++ project-fortis-interfaces/src/routes/AppPage.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import BrowserDetection from 'react-browser-detection';
 import createReactClass from 'create-react-class';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import Fluxxor from 'fluxxor';
@@ -150,11 +149,11 @@ export const AppPage = createReactClass({
 
   render() {
     if (this.shouldRenderLogin()) {
-      return this.browserDetectHandler(this.renderLogin);
+      return this.renderLogin();
     }
 
     if (this.shouldRenderUnknownCategory()) {
-      return this.browserDetectHandler(this.renderUnknownCategory);
+      return this.renderUnknownCategory();
     }
 
     if (this.shouldRenderError()) {
@@ -165,25 +164,7 @@ export const AppPage = createReactClass({
       return this.renderApp();
     }
 
-    return this.browserDetectHandler(this.renderLoading);
-  },
-
-  browserDetectHandler(renderRoute) {
-    const browserConfigs = {
-      chrome: () => renderRoute(),
-      firefox: () => renderRoute(),
-      default: () => (
-        <div className="loadingPage">
-          <h1>Your browser is not supported. Please try using Chrome or Firefox</h1>
-        </div>
-      )
-    }
-
-    return (
-      <BrowserDetection>
-        {browserConfigs}
-      </BrowserDetection>
-    )
+    return this.renderLoading();
   },
 
   renderLogin() {
