+++ project-fortis-interfaces/src/routes/routes.js
@@ -5,9 +5,30 @@ import { FactsPage } from './FactsPage';
 import { AdminPage } from './AdminPage';
 import { NotFoundPage } from './NotFoundPage';
 import React from 'react';
+import BrowserDetection from 'react-browser-detection';
+
+class AppPageWithBrowserDetection extends React.Component {
+  render() {
+    const browserConfigs = {
+      chrome: () => <AppPage {...this.props} />,
+      firefox: () => <AppPage {...this.props} />,
+      default: () => (
+        <div className="loadingPage">
+          <h1>Your browser is not supported. Please try using Chrome or Firefox</h1>
+        </div>
+      )
+    }
+
+    return (
+      <BrowserDetection>
+        { browserConfigs }
+      </BrowserDetection>
+    )
+  }
+}
 
 export const routes = (
-  <Route path="/" component={AppPage} linkLabel="My App" icon="fa fa-share-alt-square fa">
+  <Route path="/" component={AppPageWithBrowserDetection} linkLabel="My App" icon="fa fa-share-alt-square fa">
     <Route path="site/:category(/share/:sharedViewState)" component={DashboardPage} />
     <Route path="site/:category/facts" component={FactsPage} />
     <Route path="site/:category/admin" component={AdminPage} />
